Move utility functions to a separate module and add tests
